Fix: Remove index backlinks and add log file to gitignore
This commit is the final addition of backlinks to the repo and should
only be used as an example.
Since backlinks are generated, we don't need to include them.
The index page now is linked as `index.pd` for ease when uploading to a
website.

Additionally, the `log.txt` file is added to `.gitignore` to prevent it from being tracked by Git. This file is likely used for logging purposes and doesn't need to be part of the repository.