feat: add script to rename wiki files and update links
This commit introduces a new script, `smv`, to facilitate renaming wiki files and updating corresponding links within the wiki.
The script uses `fzf` to select the file to rename, prompts the user for a new name, constructs new and old linkified versions of the filename, and then uses `sed` to replace all instances of the old link with the new link in all `.pd` files.
Finally, it renames the file.

Additionally, the `build.sh` script now calls the `backlinks.sh` script.
The graph_maker.sh file now includes commented lines of shell to call it.